rename tooltip => popup
