+++ src/pages/admin/UserPage.tsx
@@ -3,8 +3,8 @@ import type { UserResponse } from '../../app/api/apiTypes/adminApiTypes';
 import validateUpdateUser from '../../components/formElements/validation/validateUpdateUser';
 import Icon from '../../components/icons/Icon';
 import useMessagePopup from '../../components/messagePopup/useMessagePopup';
+import Tooltip from '../../components/popup/Popup';
 import Table from '../../components/sortTable/Table';
-import Tooltip from '../../components/tooltip/Tooltip';
 import { useGetAllUsersQuery } from '../../features/adminUsers/adminUserApiSlice';
 import EditUserInput from '../../features/adminUsers/components/EditUserInput';
 import useAuth from '../../features/auth/hooks/useAuth';
